centralize app navigation theme
