+++ apps/tlon-mobile/src/App.main.tsx
@@ -2,8 +2,6 @@ import { useAsyncStorageDevTools } from '@dev-plugins/async-storage';
 import { useReactNavigationDevTools } from '@dev-plugins/react-navigation';
 import { useReactQueryDevTools } from '@dev-plugins/react-query';
 import {
-  DarkTheme,
-  DefaultTheme,
   NavigationContainer,
   NavigationContainerRefWithCurrent,
   NavigationState,
@@ -20,6 +18,7 @@ import { useRequiredUpdate } from '@tloncorp/app/hooks/useRequiredUpdate';
 import { useResetDb } from '@tloncorp/app/hooks/useResetDb';
 import { useMigrations } from '@tloncorp/app/lib/nativeDb';
 import { splashScreenProgress } from '@tloncorp/app/lib/splashscreen';
+import { useAppNavigationTheme } from '@tloncorp/app/navigation/useAppNavigationTheme';
 import { AppDataProvider } from '@tloncorp/app/provider/AppDataProvider';
 import { BaseProviderStack } from '@tloncorp/app/provider/BaseProviderStack';
 import {
@@ -181,7 +180,17 @@ export function ConnectedAppContent({
 }: {
   migrationState: MigrationState;
 }) {
-  const isDarkMode = useIsDarkMode();
+  return (
+    <FeatureFlagConnectedInstrumentationProvider>
+      <BaseProviderStack migrationState={migrationState}>
+        <ConnectedNavigationContent />
+      </BaseProviderStack>
+    </FeatureFlagConnectedInstrumentationProvider>
+  );
+}
+
+function ConnectedNavigationContent() {
+  const navigationTheme = useAppNavigationTheme();
   const navigationContainerRef = useNavigationContainerRef();
   const routeNameRef = useRef<string>(undefined);
   const splashIsHidden = useSplashHider();
@@ -210,31 +219,27 @@ export function ConnectedAppContent({
   };
 
   return (
-    <FeatureFlagConnectedInstrumentationProvider>
-      <NavigationContainer
-        theme={isDarkMode ? DarkTheme : DefaultTheme}
-        ref={navigationContainerRef}
-        onReady={onReady}
-        onStateChange={onStateChange}
-        navigationInChildEnabled
-      >
-        <BaseProviderStack migrationState={migrationState}>
-          <ErrorBoundary>
-            <BranchProvider>
-              <GestureHandlerRootView style={{ flex: 1 }}>
-                <SignupProvider>
-                  {splashIsHidden ? <App /> : null}
+    <NavigationContainer
+      theme={navigationTheme}
+      ref={navigationContainerRef}
+      onReady={onReady}
+      onStateChange={onStateChange}
+      navigationInChildEnabled
+    >
+      <ErrorBoundary>
+        <BranchProvider>
+          <GestureHandlerRootView style={{ flex: 1 }}>
+            <SignupProvider>
+              {splashIsHidden ? <App /> : null}
 
-                  {__DEV__ && (
-                    <DevTools navigationContainerRef={navigationContainerRef} />
-                  )}
-                </SignupProvider>
-              </GestureHandlerRootView>
-            </BranchProvider>
-          </ErrorBoundary>
-        </BaseProviderStack>
-      </NavigationContainer>
-    </FeatureFlagConnectedInstrumentationProvider>
+              {__DEV__ && (
+                <DevTools navigationContainerRef={navigationContainerRef} />
+              )}
+            </SignupProvider>
+          </GestureHandlerRootView>
+        </BranchProvider>
+      </ErrorBoundary>
+    </NavigationContainer>
   );
 }
